+++ src/Routes.js
@@ -1,14 +1,13 @@
 import React, { Component } from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import MainPage from "./pages/MainPage/MainPage";
-import StationeryList from "./pages/StationeryList/StationeryList";
 import DetailPage from "./pages/DetailPage/DetailPage";
 import Welcome from "./pages/Welcome/Welcome";
 import JoinStepOne from "./pages/JoinSteps/JoinStepOne";
 import JoinStepTwo from "./pages/JoinSteps/JoinStepTwo";
 import JoinStepThree from "./pages/JoinSteps/JoinStepThree";
 import Login from "./pages/Login/Login";
-import ProductList from "./pages/components/ProductList/ProductList";
+import ProductList from "./pages/ProductList/ProductList";
 import NavContainer from "./pages/components/NavContainer/NavContainer";
 import MyPageContainer from "./pages/MyPage/MyPageContainer";
 import Footer from "./pages/components/Footer/Footer";
@@ -19,7 +18,6 @@ class Routes extends Component {
       <Router>
         <NavContainer />
         <Switch>
-          <Route exact path="/StationeryList" component={StationeryList} />
           <Route exact path="/detail/:id" component={DetailPage} />
           <Route exact path="/Welcome" component={Welcome} />
           <Route exact path="/products" component={ProductList} />
@@ -30,6 +28,7 @@ class Routes extends Component {
           <Route exact path="/" component={MainPage} />
           <Route exact path="/detail/:id" component={DetailPage} />
           <Route exact path="/MyPage" component={MyPageContainer} />
+          <Route exact path="/ProductList" component={ProductList} />
         </Switch>
         <Footer />
       </Router>
