+++ src/index.js
@@ -4,9 +4,4 @@ import Routes from "./Routes";
 import "./Styles/reset.scss";
 import "./Styles/common.scss";
 
-ReactDOM.render(
-  <>
-    <Routes />
-  </>,
-  document.getElementById("root")
-);
+ReactDOM.render(<Routes />, document.getElementById("root"));
