+++ src/pages/ProductList/component/ProductListNav/ProductListNav.scss
@@ -1,10 +1,9 @@
 .StationeryList {
-    padding: 0 20px 0px 20px;
+    padding: 100px 20px 0px 20px;
     display: flex;
     justify-content: center;
     .inner{
         display: flex;
-        width: 100%;
         height: auto;
         .categoryArea {
             position: relative;
@@ -33,29 +32,31 @@
                     width: 224px;
                     justify-content: center;
                     float: left;
-                    .depth2Title {
+
+                    span {
                         display: block;
                         font-size: 16px;
                         font-weight: 800;
-                        margin-bottom: 15px;
                         color: #000000;
+                        cursor: pointer;
                     }
                 
                     ul {
+                        display: flex;
+                        flex-direction: column;
                         padding-top: 10px;
+                        width: 224px;
+
                         .depth3 {
                             font-size: 14px;
                             line-height: 24px;
                             text-decoration: none;
                             color: #555555;
-                        }
-                        .depth3:hover {
-                            border-bottom :1px solid #000000;
-                            color: #000000;
-                        }
-                            li {
-                            float: left;
-                            width: 200px;   
+                            cursor: pointer;
+
+                            &:hover {
+                            text-decoration: underline;
+                            }
                         }
                     }
                 }
