--- src/pages/StationeryList/StationeryList.js
@@ -1,12 +1,0 @@
-import React, { Component } from "react";
-import ProductList from "../components/ProductList/ProductList";
-
-export default class StationeryList extends Component {
-  render() {
-    return (
-      <div>
-        <ProductList />
-      </div>
-    );
-  }
-}
